refactor: [#1228] move statistics back from tracker-core to main lib
The statistics are only used at the higher levels: UDP and HTTP tracker.

We will move them to new packages.
