+++ packages/tracker-core/Cargo.toml
@@ -20,7 +20,6 @@ bittorrent-http-protocol = { version = "3.0.0-develop", path = "../http-protocol
 bittorrent-primitives = "0.1.0"
 chrono = { version = "0", default-features = false, features = ["clock"] }
 derive_more = { version = "1", features = ["as_ref", "constructor", "from"] }
-futures = "0"
 r2d2 = "0"
 r2d2_mysql = "25"
 r2d2_sqlite = { version = "0", features = ["bundled"] }
