+++ src/packages/statistics/event/listener.rs
@@ -2,7 +2,7 @@ use tokio::sync::mpsc;
 
 use super::handler::handle_event;
 use super::Event;
-use crate::statistics::repository::Repository;
+use crate::packages::statistics::repository::Repository;
 
 pub async fn dispatch_events(mut receiver: mpsc::Receiver<Event>, stats_repository: Repository) {
     while let Some(event) = receiver.recv().await {
