+++ src/packages/statistics/event/sender.rs
@@ -13,10 +13,10 @@ pub trait Sender: Sync + Send {
     fn send_event(&self, event: Event) -> BoxFuture<'_, Option<Result<(), SendError<Event>>>>;
 }
 
-/// An [`statistics::EventSender`](crate::core::statistics::event::sender::Sender) implementation.
+/// An [`statistics::EventSender`](crate::packages::statistics::event::sender::Sender) implementation.
 ///
 /// It uses a channel sender to send the statistic events. The channel is created by a
-/// [`statistics::Keeper`](crate::core::statistics::keeper::Keeper)
+/// [`statistics::Keeper`](crate::packages::statistics::keeper::Keeper)
 #[allow(clippy::module_name_repetitions)]
 pub struct ChannelSender {
     pub(crate) sender: mpsc::Sender<Event>,
