+++ src/packages/statistics/keeper.rs
@@ -51,9 +51,9 @@ impl Keeper {
 
 #[cfg(test)]
 mod tests {
-    use crate::statistics::event::Event;
-    use crate::statistics::keeper::Keeper;
-    use crate::statistics::metrics::Metrics;
+    use crate::packages::statistics::event::Event;
+    use crate::packages::statistics::keeper::Keeper;
+    use crate::packages::statistics::metrics::Metrics;
 
     #[tokio::test]
     async fn should_contain_the_tracker_statistics() {
