+++ src/packages/statistics/mod.rs
@@ -1,0 +1,6 @@
+pub mod event;
+pub mod keeper;
+pub mod metrics;
+pub mod repository;
+pub mod services;
+pub mod setup;
