+++ src/packages/statistics/services.rs
@@ -2,14 +2,14 @@
 //!
 //! It includes:
 //!
-//! - A [`factory`](crate::statistics::setup::factory) function to build the structs needed to collect the tracker metrics.
-//! - A [`get_metrics`] service to get the tracker [`metrics`](crate::core::statistics::metrics::Metrics).
+//! - A [`factory`](crate::packages::statistics::setup::factory) function to build the structs needed to collect the tracker metrics.
+//! - A [`get_metrics`] service to get the tracker [`metrics`](crate::packages::statistics::metrics::Metrics).
 //!
 //! Tracker metrics are collected using a Publisher-Subscribe pattern.
 //!
 //! The factory function builds two structs:
 //!
-//! - An statistics event [`Sender`](crate::core::statistics::event::sender::Sender)
+//! - An statistics event [`Sender`](crate::packages::statistics::event::sender::Sender)
 //! - An statistics [`Repository`]
 //!
 //! ```text
@@ -21,7 +21,7 @@
 //! There is an event listener that is receiving all the events and processing them with an event handler.
 //! Then, the event handler updates the metrics depending on the received event.
 //!
-//! For example, if you send the event [`Event::Udp4Connect`](crate::core::statistics::event::Event::Udp4Connect):
+//! For example, if you send the event [`Event::Udp4Connect`](crate::packages::statistics::event::Event::Udp4Connect):
 //!
 //! ```text
 //! let result = event_sender.send_event(Event::Udp4Connect).await;
@@ -38,12 +38,13 @@
 //! ```
 use std::sync::Arc;
 
-use bittorrent_tracker_core::statistics::metrics::Metrics;
-use bittorrent_tracker_core::statistics::repository::Repository;
 use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
+use packages::statistics::metrics::Metrics;
+use packages::statistics::repository::Repository;
 use tokio::sync::RwLock;
 use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
 
+use crate::packages;
 use crate::servers::udp::server::banning::BanService;
 
 /// All the metrics collected by the tracker.
@@ -111,13 +112,14 @@ mod tests {
     use std::sync::Arc;
 
     use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
-    use bittorrent_tracker_core::{self, statistics};
+    use bittorrent_tracker_core::{self};
     use tokio::sync::RwLock;
     use torrust_tracker_configuration::Configuration;
     use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
     use torrust_tracker_test_helpers::configuration;
 
-    use crate::core::statistics::services::{get_metrics, TrackerMetrics};
+    use crate::packages::statistics;
+    use crate::packages::statistics::services::{get_metrics, TrackerMetrics};
     use crate::servers::udp::server::banning::BanService;
     use crate::servers::udp::server::launcher::MAX_CONNECTION_ID_ERRORS_PER_IP;
