+++ src/servers/apis/v1/context/stats/resources.rs
@@ -2,7 +2,7 @@
 //! API context.
 use serde::{Deserialize, Serialize};
 
-use crate::core::statistics::services::TrackerMetrics;
+use crate::packages::statistics::services::TrackerMetrics;
 
 /// It contains all the statistics generated by the tracker.
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
@@ -118,11 +118,12 @@ impl From<TrackerMetrics> for Stats {
 
 #[cfg(test)]
 mod tests {
-    use bittorrent_tracker_core::statistics::metrics::Metrics;
+    use packages::statistics::metrics::Metrics;
     use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
 
     use super::Stats;
-    use crate::core::statistics::services::TrackerMetrics;
+    use crate::packages::statistics::services::TrackerMetrics;
+    use crate::packages::{self};
 
     #[test]
     fn stats_resource_should_be_converted_from_tracker_metrics() {
