+++ src/servers/udp/server/launcher.rs
@@ -3,9 +3,9 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use bittorrent_tracker_client::udp::client::check;
-use bittorrent_tracker_core::statistics;
 use derive_more::Constructor;
 use futures_util::StreamExt;
+use packages::statistics;
 use tokio::select;
 use tokio::sync::oneshot;
 use tokio::time::interval;
@@ -14,6 +14,7 @@ use tracing::instrument;
 use super::request_buffer::ActiveRequests;
 use crate::bootstrap::jobs::Started;
 use crate::container::UdpTrackerContainer;
+use crate::packages;
 use crate::servers::logging::STARTED_ON;
 use crate::servers::registar::ServiceHealthCheckJob;
 use crate::servers::signals::{shutdown_signal_with_message, Halted};
